+++ Assets/Scenes/Overworld.unity
@@ -1084,6 +1084,8 @@ GameObject:
   - component: {fileID: 550432077}
   - component: {fileID: 550432076}
   - component: {fileID: 550432075}
+  - component: {fileID: 550432079}
+  - component: {fileID: 550432080}
   m_Layer: 5
   m_Name: Canvas
   m_TagString: Untagged
@@ -1176,6 +1178,33 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0, y: 0}
+--- !u!114 &550432079
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 550432074}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 40ac524d22b415e4281f1b8331f2b50f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &550432080
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 550432074}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 13336e05d396b874dbc94e53c2c23653, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  player: {fileID: 260865020}
+  endScreen: {fileID: 1882102159}
+  txt: {fileID: 2130763529}
 --- !u!1 &603726965
 GameObject:
   m_ObjectHideFlags: 0
@@ -2615,6 +2644,31 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 550432078}
     m_Modifications:
+    - target: {fileID: 1605876452642516558, guid: 63806c320f1b20447ba63ab9c15754f9,
+        type: 3}
+      propertyPath: player
+      value: 
+      objectReference: {fileID: 260865020}
+    - target: {fileID: 1844933589364983014, guid: 63806c320f1b20447ba63ab9c15754f9,
+        type: 3}
+      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_Target
+      value: 
+      objectReference: {fileID: 1882102160}
+    - target: {fileID: 1844933589364983014, guid: 63806c320f1b20447ba63ab9c15754f9,
+        type: 3}
+      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_MethodName
+      value: MenuReturn
+      objectReference: {fileID: 0}
+    - target: {fileID: 1844933589530272255, guid: 63806c320f1b20447ba63ab9c15754f9,
+        type: 3}
+      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_Target
+      value: 
+      objectReference: {fileID: 1882102160}
+    - target: {fileID: 1844933589530272255, guid: 63806c320f1b20447ba63ab9c15754f9,
+        type: 3}
+      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_MethodName
+      value: CloseGame
+      objectReference: {fileID: 0}
     - target: {fileID: 1844933590172177422, guid: 63806c320f1b20447ba63ab9c15754f9,
         type: 3}
       propertyPath: m_Pivot.x
@@ -2725,16 +2779,21 @@ PrefabInstance:
       propertyPath: m_Name
       value: End Game
       objectReference: {fileID: 0}
-    - target: {fileID: 1844933590172177425, guid: 63806c320f1b20447ba63ab9c15754f9,
+    - target: {fileID: 1844933590172177423, guid: 63806c320f1b20447ba63ab9c15754f9,
         type: 3}
-      propertyPath: txt
-      value: 
-      objectReference: {fileID: 1707619172}
-    - target: {fileID: 1844933590172177425, guid: 63806c320f1b20447ba63ab9c15754f9,
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1844933590565223438, guid: 63806c320f1b20447ba63ab9c15754f9,
         type: 3}
-      propertyPath: player
-      value: 
-      objectReference: {fileID: 260865020}
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3716370843111040549, guid: 63806c320f1b20447ba63ab9c15754f9,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 3
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 63806c320f1b20447ba63ab9c15754f9, type: 3}
 --- !u!1 &1449655384
@@ -3017,12 +3076,6 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1611705563}
   m_CullTransparentMesh: 0
---- !u!1 &1707619172 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 1844933590565223438, guid: 63806c320f1b20447ba63ab9c15754f9,
-    type: 3}
-  m_PrefabInstance: {fileID: 1379985834}
-  m_PrefabAsset: {fileID: 0}
 --- !u!1 &1772265402
 GameObject:
   m_ObjectHideFlags: 0
@@ -3251,6 +3304,18 @@ GameObject:
     type: 3}
   m_PrefabInstance: {fileID: 1379985834}
   m_PrefabAsset: {fileID: 0}
+--- !u!114 &1882102160 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 1605876452642516558, guid: 63806c320f1b20447ba63ab9c15754f9,
+    type: 3}
+  m_PrefabInstance: {fileID: 1379985834}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1882102159}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 40ac524d22b415e4281f1b8331f2b50f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!224 &1882102161 stripped
 RectTransform:
   m_CorrespondingSourceObject: {fileID: 1844933590172177422, guid: 63806c320f1b20447ba63ab9c15754f9,
@@ -3560,3 +3625,15 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 7065029016f68bb4e9ec76b9d7f097e0, type: 3}
+--- !u!114 &2130763529 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 4300561032611320231, guid: 63806c320f1b20447ba63ab9c15754f9,
+    type: 3}
+  m_PrefabInstance: {fileID: 1379985834}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5f7201a12d95ffc409449d95f23cf332, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
