+++ Assets/Scripts/EndGame.cs
@@ -4,26 +4,10 @@ using UnityEngine;
 using UnityEngine.SceneManagement;
 using TopdownCharacterController;
 using UnityEngine.UI;
+using UnityEngine.UIElements;
 
 public class EndGame : MonoBehaviour
 {
-
-    public GameObject player;
-    public GameObject txt;
-
-    public void Start()
-    {
-        Health health = player.GetComponent<Health>();
-        health.DeathCallbacks.Add(DeathCallback);
-    }
-
-    private void DeathCallback(Dictionary<string, string> damageInfo)
-    {
-        Time.timeScale = 0;
-        txt.GetComponent<Text>().text = "You Died!";
-        gameObject.SetActive(true);
-    }
-
     public void MenuReturn()
     {
         Time.timeScale = 1;
