+++ Assets/Scripts/EndScreenCheck.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 13336e05d396b874dbc94e53c2c23653
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
